+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Added related works project overview page [#700]
 
 ## Updated
+- Updated Research Output standardFields config so that only `title` field is required when answering question [#33]
 - `Custom Question` updates [#929]
 - Moved `addQuestionMutation` out of `QuestionView` so that `QuestionView` can be shared for both `BASE` and `CUSTOM` questions
   - Updated `QuestionTypeSelectPage` to include the `addQuestionMutation` and `getDisplayOrder` and pass those to `QuestionAdd`